Adding references for literature search
A Forensic Email Analysis Tool Using Dynamic Visualization article focuses on the development and experimentation of a tool that they created for email analysis similar to EMT. They provide a framework of what results they expected to yield from their tool. Although they do not offer any glimpse of the code that they created, it provides perspective in regards to their strengths and shortcomings. 

The article "A System for Email Recipient Prediction" from the book From Social Data Mining and Analysis to Prediction and Community Detection looks into increasing the accuracy of email recipient auto-filling and attachment filtering in email programs with learning behaviors of the individual through their email history.  

"Profiling individuals based on email analysis and ego networks: A visualization technique" article also deals with profiling an individual's emails and creating/identifying the ego network in which the individual is related to.